feat:camel 2 to camel 3 and Junit 4 to Junit 5 migration(RDCC-2493)
