+++ src/main/java/uk/gov/hmcts/reform/juddata/camel/processor/JudicialOfficeAppointmentProcessor.java
@@ -85,8 +85,6 @@ public class JudicialOfficeAppointmentProcessor
             invalidJudicialUserProfileRecords, judicialOfficeAppointmentJsrValidatorInitializer, exchange,
             logComponentName);
 
-        log.info("{}:: Judicial Appointment Records count after Validation {}:: ", logComponentName,
-            filteredJudicialAppointments.size());
 
         filterAppointmentsRecordsForForeignKeyViolation(filteredJudicialAppointments, exchange);
 
@@ -95,6 +93,8 @@ public class JudicialOfficeAppointmentProcessor
         if (judicialOfficeAppointments.size() != filteredJudicialAppointments.size()) {
             setFileStatus(exchange, applicationContext);
         }
+        log.info("{}:: Judicial Appointment Records count after Validation {}:: ", logComponentName,
+            filteredJudicialAppointments.size());
 
         exchange.getMessage().setBody(filteredJudicialAppointments);
     }
