+++ src/test/java/uk/gov/hmcts/reform/juddata/camel/processor/JudicialOfficeAuthorisationProcessorTest.java
@@ -5,8 +5,8 @@ import org.apache.camel.Exchange;
 import org.apache.camel.Message;
 import org.apache.camel.impl.DefaultCamelContext;
 import org.apache.camel.spi.Registry;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -74,7 +74,7 @@ public class JudicialOfficeAuthorisationProcessorTest {
     final PlatformTransactionManager platformTransactionManager = mock(PlatformTransactionManager.class);
     final TransactionStatus transactionStatus = mock(TransactionStatus.class);
 
-    @Before
+    @BeforeEach
     public void setup() {
 
         judicialOfficeAuthorisationProcessor = new JudicialOfficeAuthorisationProcessor();
