+++ src/test/java/uk/gov/hmcts/reform/juddata/camel/task/LeafRouteTaskTest.java
@@ -1,5 +1,11 @@
 package uk.gov.hmcts.reform.juddata.camel.task;
 
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.batch.repeat.RepeatStatus;
+import uk.gov.hmcts.reform.data.ingestion.camel.route.DataLoadRoute;
+import uk.gov.hmcts.reform.juddata.camel.util.JrdExecutor;
+
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
@@ -12,12 +18,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.util.ReflectionTestUtils.setField;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.springframework.batch.repeat.RepeatStatus;
-import uk.gov.hmcts.reform.data.ingestion.camel.route.DataLoadRoute;
-import uk.gov.hmcts.reform.juddata.camel.util.JrdExecutor;
-
 public class LeafRouteTaskTest {
     LeafRouteTask leafRouteTask = spy(new LeafRouteTask());
 
@@ -25,7 +25,7 @@ public class LeafRouteTaskTest {
 
     JrdExecutor jrdExecutor = mock(JrdExecutor.class);
 
-    @Before
+    @BeforeEach
     public void init() {
         setField(leafRouteTask, "logComponentName", "testlogger");
         setField(leafRouteTask, "dataLoadRoute", dataLoadRoute);
