+++ src/test/java/uk/gov/hmcts/reform/juddata/camel/task/ParentRouteTaskTest.java
@@ -23,7 +23,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.util.ReflectionTestUtils.setField;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class ParentRouteTaskTest {
 
     @InjectMocks
